Split the hidden-backlog cases out of a file over the ceiling
